Update README.md
Consider making the README a little more friendly to the first time stumbler coming across the project. For example, this README makes it very clear that you're not reading a step wise document. I had a friend once (couldn't be me obviously :P) that told me he was trying to play around with the software in a rapid and gluttonous fashion. He performed the pip and npm commands.. and then thought that the next step in getting it all setup was to build the dependencies, make the certs, etc.